Create README.md
R Programming Assignments - 1, 2and 3
======================================

Programming assignments for the course "R Programming"

**Programming Assignment 1 - Air Pollution** 
For this first programming assignment you will write three functions that are meant to  
interact with dataset that accompanies this assignment. The dataset is contained in a zip  
file specdata.zip that you can download from the Coursera web site. 

**Programming Assignment 2 - Lexical Scoping**  
This second programming assignment will require you to write an R function is able to  
cache potentially time-consuming computations. For example, taking the mean of a numeric  
vector is typically a fast operation. However, for a very long vector, it may take too  
long to compute the mean, especially if it has to be computed repeatedly (e.g. in a loop).  
If the contents of a vector are not changing, it may make sense to cache the value of the 
 mean so that when we need it again, it can be looked up in the cache rather than  
 recomputed. In this Programming Assignment will take advantage of the scoping rules of  
 the R language and how they can be manipulated to preserve state inside of an R object.
 
**Programming Assignment 3 - Hospital Quality**  
See the file "rprog-doc-ProgAssignment3.pdf" inside the Assignment 3 directory for more  
information about this assignment.